Change Month type to ByteType

+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/expressions/intervalExpressions.scala
@@ -87,7 +87,7 @@ case class Quarter(child: Expression) extends UnaryExpression with ExpectsInputT
 
 case class Month(child: Expression) extends UnaryExpression with ExpectsInputTypes {
   override def inputTypes: Seq[AbstractDataType] = Seq(CalendarIntervalType)
-  override def dataType: DataType = IntegerType
+  override def dataType: DataType = ByteType
   override protected def nullSafeEval(date: Any): Any = {
     IntervalUtils.getMonth(date.asInstanceOf[CalendarInterval])
   }
